\name{clust_cormap}
\alias{clust_cormap}
\title{Helper function to cluster the correlation matrix and return the sorted matrix for plotting.}
\usage{
  clust_cormap(cormat, minfrac = 0.1, distfn = function(cm) (1 - cm),
               method = "complete", cor.thr = 0.8, cor.mar = 0.05, cut.thr = 0.9,
               cut.size = 1, list.output = FALSE)
}
\arguments{
  \item{cormat}{\code{numeric matrix or list}. The correlation matrix or a list containing matrices of correlation values, counts and P-Values as generated by \command{eset_cor}. If a list then the correlation matrix is exprected to be in slot 'cor', the counts matrix in slot 'counts' and the matrix of P-Values in slot 'pvalues'.}

  \item{minfrac}{\code{numeric}. Percent of NAs allowed per row. Only rows and columns with less than the given fraction of missing values are retained; defaults to 10 per cent.}

  \item{distfn}{\code{function}. Function to calculate the dissimilarity matrix for clustering.}

  \item{method}{\code{character}. The agglomeration method to be used for clustering. See help for \code{\link[stats]{hclust}}.}

  \item{cor.thr}{\code{numeric}. Correlation threshold. This value is sign-sensitive: a negative threshold will retain rows and columns of the correlation matrix with correlation values between -1 and the threshold, a positive value will retain rows and columns with values between the threshold and 1. Zero (0) is treated as positive. Default is \code{NULL}, disabling filtering.}

  \item{cor.mar}{\code{numeric}. Margin of values needed meeting \code{cor.thr}. Only rows (and columns) with at least (100 * \code{cor.mar}) per cent of columns (or rows, respectively) meeting \code{cor.thr} and vice versa are retained. Defaults to 80 per cent.}

  \item{cut.thr}{\code{numeric}. Threshold for tree cutting, i.e., the height at which branches are to be cut. Passed to \code{WGCNA::cutreeStatic}.}

  \item{cut.size}{\code{numeric}. The minimum number of object on a branch to be considered a cluster. Passed to \code{WGCNA::cutreeStatic}.}

  \item{list.output}{\code{logical}. Should the output be a list of different object created in the call? Depends also on input type.}
}
\description{
  Helper function to cluster the correlation matrix and return the sorted matrix for plotting.
}

