\name{cor_maps}
\alias{cor_maps}
\title{Wrapper function to draw up to three different heatmaps from three different data sets and in addition, draw two heatmaps
  showing the difference between the correlation maps of the first tow data sets.}
\usage{
  cor_maps(full.eset.first, full.eset.second, full.eset.third = NULL,
           heading = NULL, clust = TRUE, groupnames = c("cancer", "normal",
                                                        "cell lines"), cex = 0.5, add.sig = FALSE, genes2highl = NULL)
}
\description{
  Wrapper function to draw up to three different heatmaps from three different data sets and in addition, draw two heatmaps
  showing the difference between the correlation maps of the first tow data sets.
}

