% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cormap2.R
\docType{package}
\name{coreheat}
\alias{coreheat}
\title{Draw Correlation Heatmaps}
\description{
Create correlation heatmaps from a numeric matrix. Optionally, data can be annotated using, e.g., BioMart,
    clustered and filtered by correlation and/or tree cutting.
}
\details{
\tabular{ll}{
Package: \tab coreheat\cr
Type: \tab Package\cr
Initial version: \tab 0.1-0\cr
Created: \tab 2016-08-11\cr
License: \tab GPL-3\cr
The main function to be called by end users is \command{cormap2} which is wrapper performing all necessary steps to create a heatmap.
}
}
\author{
Vidal Fey <vidal.fey@gmail.com>, Henri Sara <henri.sara@gmail.com>
Maintainer: Vidal Fey <vidal.fey@gmail.com>
}
\keyword{package}
