\name{cormap3}
\alias{cormap3}
\title{Function to automatically split clusters based on noise level and hierarchy, produce a list of vectors of genes, ...}
\usage{
  cormap3(eset, minfrac = 0.1, method = "ward", do.abs = TRUE,
          main = "correlation map", threshold = 2, cex = 0.2, h = NULL,
          cex.clust = cex, cex.filt = cex, do.plots = 1:3, genes2highl = NULL)
}
\description{
  Function to automatically split clusters based on noise level and hierarchy, produce a list of vectors of genes, ...
}

