\name{heatmap.cor}
\alias{heatmap.cor}
\title{Helper function to draw the heatmap. Wrapper around 'heatmap.n2' from the 'heatmapGen2' package.}
\usage{
  heatmap.cor(cormat, order_list = TRUE, main = "",
              main_postfix = "Dissimilarity = 1 - Correlation", x.labels = NULL,
              y.labels = NULL, labelheight = 0.2, labelwidth = 0.2, cex = 0.5,
              add.sig = FALSE, pv = NULL, genes2highl = NULL)
}
\description{
  Helper function to draw the heatmap. Wrapper around 'heatmap.n2' from the 'heatmapGen2' package.
}

