\docType{package}
\name{coreheat}
\alias{coreheat}
\alias{coreheat-package}
\title{Draw correlation heatmaps.}
\description{
  Create correlation heatmaps from a numeric matrix. Optionally, data can be annotated using, e.g., biomart, clustered and filtered by correlation and/or tree cutting.
}
\details{
  \tabular{ll}{
    Package: \tab coreheat\cr
    Type: \tab Package\cr
    Initial version: \tab 0.1-0\cr
    License: \tab To be used with CORPLOT shiny application only. Please contact the authors for more information.\cr
  }
  The main function to be called by end users is \command{cormap2} which is wrapper performing all necessary steps to create a heatmap.
}
\note{
  Based on code created by Henri Sara
}
\keyword{package}

