% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cormap2.R
\docType{package}
\name{coreheat}
\alias{coreheat}
\title{Draw Correlation Heatmaps}
\description{
Create correlation heatmaps from a numeric matrix. Ensembl Gene ID row names can be converted to Gene Symbols
    using, e.g., BioMart. Optionally, data can be clustered and filtered by correlation, tree cutting and/or number
    of missing values. Genes of interest can be highlighted in the plot and correlation significance be indicated by
    asterisks encoding corresponding P-Values. Plot dimensions and label measures are adjusted automatically by default.
    The plot features rely on the heatmap.n2() function in the 'heatmapFlex' package.
}
\details{
\tabular{ll}{
Package: \tab coreheat\cr
Type: \tab Package\cr
Initial version: \tab 0.1-0\cr
Created: \tab 2016-08-11\cr
License: \tab GPL-3\cr
The main function to be called by end users is \command{cormap2} which is wrapper performing all necessary steps to create a heatmap.
}
}
\author{
Vidal Fey <vidal.fey@gmail.com>, Henri Sara <henri.sara@gmail.com>
Maintainer: Vidal Fey <vidal.fey@gmail.com>
}
\keyword{package}
